+++ src/main.cpp
@@ -17,6 +17,12 @@ using json = nlohmann::json;
 namespace Eternal {
 	json data;
 	bool canUseData;
+	std::string level_name;
+	std::string level_date;
+	std::string level_creators;
+	int level_creator_count;
+	int level_top;
+	bool is_eternal_level;
 }
 
 $on_mod(Loaded){
@@ -115,21 +121,14 @@ class $modify(EternalProfilePage, ProfilePage){
 
 class $modify(EternalLevelInfoLayer, LevelInfoLayer){
 
-	struct Fields {
-		std::string level_name;
-		std::string level_date;
-		std::string level_creators;
-		int top;
-		bool is_eternal_level;
-	};
-
 	bool init(GJGameLevel* level, bool challange){
 		if(!LevelInfoLayer::init(level, challange)){
 			return false;
 		}
-		m_fields->is_eternal_level = false;
+		Eternal::is_eternal_level = false;
 
 		if(Eternal::canUseData){
+			// Find a level in the demonlist
 			json demonlist = Eternal::data["demonlist"];
 			for(int i=0; i<demonlist.size(); i++){
 				std::string level_name = demonlist[i][0];
@@ -137,31 +136,62 @@ class $modify(EternalLevelInfoLayer, LevelInfoLayer){
 				std::string level_date = demonlist[i][2];
 
 				if(level_name == level->m_levelName){
-					m_fields->level_name = level_name;
-					m_fields->level_creators = level_creators;
-					m_fields->level_date = level_date;
-					m_fields->top = i+1;
-					m_fields->is_eternal_level = true;
+					Eternal::level_name = level_name;
+					Eternal::level_creators = level_creators;
+					Eternal::level_date = level_date;
+					Eternal::level_top = i+1;
+					Eternal::is_eternal_level = true;
+					Eternal::level_creator_count = 0;
+
+					for(int j=0; j<Eternal::level_creators.length(); j++){
+						if(Eternal::level_creators.at(j) == ','){
+							Eternal::level_creator_count++;
+						}
+					}
+					break;
 				}
 			}
-			if(!m_fields->is_eternal_level){
+			if(!Eternal::is_eternal_level){
 				return true;
 			}
 
+			// Create a menu for the top label (otherwise we wouldn't be able to click it)
 			auto EternalTopMenu = CCMenu::create();
 			EternalTopMenu->setPosition({151, 171});
 			EternalTopMenu->setContentSize({64, 13});
 			EternalTopMenu->setID("eternal-top-menu");
 
-			auto label_spr = CCLabelBMFont::create(std::format("Top #{}", m_fields->top).c_str(), "chatFont.fnt");
+			// Create a top label
+			auto label_spr = CCLabelBMFont::create(std::format("Top #{}", Eternal::level_top).c_str(), "chatFont.fnt");
 
 			auto label = CCMenuItemSpriteExtra::create(label_spr, EternalTopMenu, menu_selector(EternalLevelInfoLayer::onTopLabelClicked));
 			label->setPosition({33, 9});
 			label->setScale(0.6);
 			label->setID("top-label");
 			
+			// Add these things
 			EternalTopMenu->addChild(label);
 			this->addChild(EternalTopMenu);
+
+			// Modify the creator name to match Eternal::level_creators
+			// pos 56.100 10.400
+			// scale 0.8
+			CCMenuItemSpriteExtra* creator_name = (CCMenuItemSpriteExtra*)this->getChildByIDRecursive("creator-name");
+			auto creator_label_pos = ((CCLabelBMFont*)creator_name->getChildren()->firstObject())->getPosition();
+			
+			auto creator_label_spr = CCLabelBMFont::create(std::format("By {}", Eternal::level_creators).c_str(), "goldFont.fnt");
+			creator_label_spr->setPosition(creator_label_pos);
+			creator_label_spr->setScale(0.8);
+
+			if(Eternal::level_creator_count > 0){
+				creator_label_spr->setScale(0.65);
+				this->removeChildByID("copy-indicator");	
+				this->removeChildByID("high-object-indicator");	
+			}
+
+			creator_name->removeAllChildren();
+			creator_name->addChild(creator_label_spr);
+
 		}
 
 		return true;
@@ -169,12 +199,10 @@ class $modify(EternalLevelInfoLayer, LevelInfoLayer){
 
 	void onTopLabelClicked(CCObject *sender){
 		this->getChildByIDRecursive("top-label")->setScale(0.6);
-		if(m_fields->is_eternal_level){
-			FLAlertLayer::create(
-				m_fields->level_name.c_str(),
-				std::format("Top: <cr>{}</c>\nClaimed it's top spot at: <cg>{}</c>\nCreators: <co>{}</c>", m_fields->top, m_fields->level_date, m_fields->level_creators),
-				"OK"
-			)->show();
-		}
+		FLAlertLayer::create(
+			Eternal::level_name.c_str(),
+			std::format("Top: <cr>{}</c>\nClaimed it's top spot at: <cg>{}</c>\nCreators: <co>{}</c>", Eternal::level_top, Eternal::level_date, Eternal::level_creators),
+			"OK"
+		)->show();
 	}
 };